+++ R/reporter.R
@@ -83,3 +83,14 @@ Reporter <- R6::R6Class("Reporter",
     }
   )
 )
+
+#' Retrieve the default reporter.
+#'
+#' This will be \code{\link{SummaryReporter}} unless option
+#' \code{testthat.default_reporter} is set
+#'
+#' @export
+#' @keywords internal
+default_reporter <- function() {
+  getOption("testthat.default_reporter", "summary")
+}